+++ src/main/java/algo/project/Graph.java
@@ -18,7 +18,7 @@ public class Graph {
     }
 
     // Method to add an edge between two vertices
-    public void addEdge(Vertex fromVertex, Vertex toVertex, int weight) {
+    public void addEdge(Vertex fromVertex, Vertex toVertex, double weight) {
         // Ensure both vertices are present in the adjacency list
         addVertex(fromVertex);
         addVertex(toVertex);
@@ -63,78 +63,3 @@ public class Graph {
         return sb.toString();
     }
 }
-
-
-// import java.util.ArrayList;
-// import java.util.HashMap;
-// import java.util.List;
-// import java.util.Map;
-
-// public class Graph {
-//     private Map<Vertex, List<Edge>> adjacencyList;
-
-//     public Graph() {
-//         this.adjacencyList = new HashMap<>();
-//     }
-
-//     public void addVertex(Vertex vertex) {
-//         adjacencyList.putIfAbsent(vertex, new ArrayList<>());
-//     }
-
-//     public void addEdge(Vertex fromVertex, Vertex toVertex, int weight) {
-//         addVertex(fromVertex);
-//         addVertex(toVertex);
-//         adjacencyList.get(fromVertex).add(new Edge(fromVertex, toVertex, weight));
-//     }
-
-//     public List<Edge> getEdges(Vertex vertex) {
-//         return adjacencyList.get(vertex);
-//     }
-
-//     public Map<Vertex, List<Edge>> getAdjacencyList() {
-//         return adjacencyList;
-//     }
-
-//     @Override
-//     public String toString() {
-//         StringBuilder sb = new StringBuilder();
-//         for (Vertex vertex : adjacencyList.keySet()) {
-//             sb.append(vertex).append(": ");
-//             for (Edge edge : adjacencyList.get(vertex)) {
-//                 sb.append(edge).append(", ");
-//             }
-//             sb.append("\n");
-//         }
-//         return sb.toString();
-//     }
-
-//     public static void main(String[] args) {
-//         Graph graph = new Graph();
-
-//         // Creating vertices
-//         Vertex garage = new Vertex(0, "Garage");
-//         Vertex pickupA = new Vertex(1, "Pick-up A");
-//         Vertex deliverA = new Vertex(2, "Deliver A");
-//         Vertex pickupB = new Vertex(3, "Pick-up B");
-//         Vertex deliverB = new Vertex(4, "Deliver B");
-
-//         // Adding edges
-//         graph.addEdge(garage, pickupA, 10);
-//         graph.addEdge(pickupA, deliverA, 20);
-//         graph.addEdge(deliverA, pickupB, 30);
-//         graph.addEdge(pickupB, deliverB, 40);
-//         graph.addEdge(deliverB, garage, 50);
-
-//         // Adding zero weight edges to complete the graph
-//         List<Vertex> vertices = List.of(garage, pickupA, deliverA, pickupB, deliverB);
-//         for (Vertex from : vertices) {
-//             for (Vertex to : vertices) {
-//                 if (from != to && graph.getEdges(from).stream().noneMatch(edge -> edge.getToVertex().equals(to))) {
-//                     graph.addEdge(from, to, 0);
-//                 }
-//             }
-//         }
-
-//         System.out.println(graph);
-//     }
-// }
